+++ textures/texname.lua
@@ -161,6 +161,7 @@ local availableTextures = {
   ["serendipity"] = true,
   ["shooting_stars"] = true,
   ["sudden_death"] = true,
+  ["sudden_doom"] = true,
   ["surge_of_light"] = true,
   ["sword_and_board"] = true,
   ["tooth_and_claw"] = true,