+++ changelog.md
@@ -6,8 +6,12 @@ The Death Knight class experiments the upcoming rework for Cataclysm.
 Make sure to report issues you would encounter, thank you :)
 
 New effects:
+- New SAO: Death Knight's Sudden Doom (Cataclysm)
 - New SAO: Hunter's Lock and Load (Season of Discovery)
 - New SAO: Warrior's Sword and Board (Season of Discovery)
+- New GAB: Death Knight's Death Coil, during Sudden Doom (Cataclysm)
+- New GAB: Death Knight's Icy Touch, during Rime (Cataclysm)
+- New GAB: Death Knight's Obliterate, during Killing Machine (Cataclysm)
 - New GAB: Paladin's Holy Shock, as combat-only counter
 - New GAB: Warrior's Overpower, during Taste for Blood* (SoD) 
 - New GAB: Warrior's Slam, during Sword and Board (SoD)